pyproject.toml
[project]
name = "open-notebook"
version = "1.8.5"
description = "An open source implementation of a research assistant, inspired by Google Notebook LM"
authors = [
    {name = "Luis Novo", email = "lfnovo@gmail.com"}
]
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.9.2",
    "loguru>=0.7.2",
    "langchain>=1.2.0",
    "langgraph>=1.0.5",
    "tiktoken>=0.12.0",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langchain-anthropic>=1.3.0",
    "langchain-ollama>=1.0.1",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langchain_mistralai>=1.1.1",
    "langchain_deepseek>=1.0.0",
    "tomli>=2.0.2",
    "python-dotenv>=1.0.1",
    "httpx[socks]>=0.27.0",
    "content-core>=1.14.1,<2",
    "ai-prompter>=0.4,<1",
    "esperanto>=2.20.0,<3",
    "surrealdb>=1.0.4",
    "podcast-creator>=0.12.0,<1",
    "surreal-commands>=1.3.1,<2",
    "numpy>=2.4.1",
    "pycountry>=26.2.16",
    "babel>=2.18.0",
    "matplotlib>=3.10.8",
    "genanki>=0.13.1",
    "graphviz>=0.21",
    "python-pptx>=1.0.2",
    "python-docx>=1.2.0",
    "Pillow>=11.0.0",
    "weasyprint>=63.0",
    "cairosvg>=2.7.1",
    "reportlab>=4.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "moviepy>=2.1.1",
    "ffmpeg-python>=0.2.0",
    "imagehash>=4.3.2",
    "tenacity>=9.0.0",
    "jinja2>=3.1.4",
]

[tool.setuptools]
package-dir = {"open_notebook" = "open_notebook"}

[tool.setuptools.package-data]
"open_notebook.artifacts.assets.fonts" = ["*.ttf", "*.otf", "LICENSE.txt"]
"open_notebook.artifacts.assets.templates" = ["*.j2", "*.svg", "*.html"]


[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.5.5",
    "mypy>=1.11.1",
    "types-requests>=2.32.0.20241016",
    "ipywidgets>=8.1.5",
    "pre-commit>=4.0.1",
    "pytest>=9.0.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.13",
    "types-requests>=2.32.4.20250913",
]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = [
    "E501",  # line too long
    "E402",  # module level import not at top of file (Streamlit requires this pattern)
    "E722",  # do not use bare except (legacy code pattern)
    "F401",  # imported but unused (may be used in type hints or re-exports)
    "F541",  # f-string without placeholders
    "F841",  # local variable assigned but never used
]

[tool.ruff.lint.per-file-ignores]
# Streamlit files need nest_asyncio.apply() before imports
"app_home.py" = ["E402"]
"pages/**/*.py" = ["E402"]

[tool.mypy]
# Exclude Streamlit UI pages from type checking
[[tool.mypy.overrides]]
module = "pages.*"
ignore_errors = true

[tool.pytest.ini_options]
markers = [
    "unit: offline tests, no LLM, no network",
    "integration: real-LLM tests, single provider, small fixtures",
    "real_llm: any test that hits a real model provider",
    "e2e: HTTP router -> command -> renderer end-to-end",
    "eval: LLM-judge rubric scoring (nightly only)",
    "matrix: cross-provider cell in the pre-merge matrix",
    "slow: test takes >15s",
]
asyncio_mode = "auto"
